html/interaction/focus/cross-frame-focus.html

<!doctype html>
<html>
  <head>
    <meta charset=utf-8>
    <title>Document.activeElement values between browsing contexts.</title>
    <link rel="help" href="https://html.spec.whatwg.org/multipage/interaction.html#focus">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
      'use strict';
      async_test(function (t) {
        var i = document.createElement('iframe');
        i.src = "//www1.web-platform.test/html/interaction/focus/iframe.html";
        var i2 = document.createElement('iframe');
        i2.src = 'about:blank';

        var waitBlur = t.step_func_done(function(e) {
          assert_array_equals(e.data, ['blur', 'BODY']);
        });
        var waitFocus = t.step_func(function(e) {
          assert_array_equals(e.data, ['focus', 'INPUT']);
          window.removeEventListener('message', waitFocus);
          window.addEventListener('message', waitBlur);
          i2.focus();
        });


        window.addEventListener('message', waitFocus);

        document.body.appendChild(i);
        document.body.appendChild(i2);
      }, "An iframe that loses focus to a sibling should clear activeElement.");
    </script>
  </body>
</html>
